update check repodirectory for exists for TestAltLinIgniteTests2xJdk811_BuildApacheIgnite

+++ .teamcity/TestAltLinIgniteTests2xJdk811/buildTypes/TestAltLinIgniteTests2xJdk811_BuildApacheIgnite.kt
@@ -96,7 +96,12 @@ object TestAltLinIgniteTests2xJdk811_BuildApacheIgnite : BuildType({
                 set -x
                 
                 mkdir -pv repository
-                cp -rfv ~/.m2/repository/org/apache/ignite/* repository/
+                mkdir -pv repository
+                if [ -d `readlink -f `~/.m2/repository/com/sbt/ignite/`]; then
+                    cp -rfv ~/.m2/repository/com/sbt/ignite/* repository/
+                else
+                    cp -rfv ~/.m2/repository/org/apache/ignite/* repository/ 
+                fi
             """.trimIndent()
         }
     }
